+++ payment-engine/src/main/java/io/gomobi/payment/repository/QrCallbackAuditRepository.java
@@ -3,6 +3,8 @@ package io.gomobi.payment.repository;
 import io.gomobi.payment.entity.QrCallbackAudit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QrCallbackAuditRepository extends JpaRepository<QrCallbackAudit, Long> {
+import java.math.BigInteger;
+
+public interface QrCallbackAuditRepository extends JpaRepository<QrCallbackAudit, BigInteger> {
 }
